+++ src/BitsToValue.cpp
@@ -271,9 +271,9 @@ struct BitsToValueModuleWidget : InfNoiseModuleWidget {
         addParam(createParamCentered<RoundSmallBlackKnob>(Vec(29.325f, 288.792f), module, BitsToValueModule::MINCNTRMAX_PARAM));
         addParam(createParamCentered<Trimpot>(Vec(59.325f, 288.792f), module, BitsToValueModule::MINCNTRMAX_TRIM_PARAM));
         addInput(createInputCentered<ThemedPJ301MPort>(Vec(89.325f, 288.792f), module, BitsToValueModule::MINCNTRMAX_INPUT));
-        addChild(createLightCentered<SmallLight<GreenLight>>(Vec(38.435f, 264.538f), module, BitsToValueModule::MIN_LIGHT));
-        addChild(createLightCentered<SmallLight<GreenLight>>(Vec(60.319f, 264.538f), module, BitsToValueModule::CNTR_LIGHT));
-        addChild(createLightCentered<SmallLight<GreenLight>>(Vec(82.713f, 264.538f), module, BitsToValueModule::MAX_LIGHT));
+        addChild(createLightCentered<TinyLight<GreenLight>>(Vec(38.435f, 264.538f), module, BitsToValueModule::MIN_LIGHT));
+        addChild(createLightCentered<TinyLight<GreenLight>>(Vec(60.319f, 264.538f), module, BitsToValueModule::CNTR_LIGHT));
+        addChild(createLightCentered<TinyLight<GreenLight>>(Vec(82.713f, 264.538f), module, BitsToValueModule::MAX_LIGHT));
         addParam(createParamCentered<infNoiseLtSmallButton<bc_green, true>>(Vec(94.450f, 269.019f), module, BitsToValueModule::MINCNTRMAX_BTN_PARAM));
 
         // Outputs