+++ src/plugin.cpp
@@ -239,6 +239,7 @@ void init(Plugin *p) {
 
 	// Add modules here
 	p->addModel(modelADREnvelope);
+	p->addModel(modelADSDREnvelope);
 	p->addModel(modelArm3XY);
 	p->addModel(modelAutoScale4);
 	p->addModel(modelBernoulliSwitch);