+++ src/plugin.hpp
@@ -13,6 +13,7 @@ extern Plugin* pluginInstance;
 // Declare each Model, defined in each module source file
 
 extern Model* modelADREnvelope;
+extern Model* modelADSDREnvelope;
 extern Model* modelArm3XY;
 extern Model* modelAutoScale4;
 extern Model* modelBernoulliSwitch;