+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 2.0.4 (In development)
+Changed the order of the modules as they appear in the browser and manual for a few modules, as there is now a "Random" and "Envelope" section in the manual, containing some of the modules there were otherwise located in the "Misc" section at the end.
+
+## New modules
+- **[ADSDR Envelope](doc/Envelope.md#adsdr-envelope)**: Attack-Decay-Sustain-Delay-Release Envelope with attack/dectay/release time and shape, and levels for Attack, Decay, Sustain and Release. Times can be affected by Rate Chaos (set via context-menu).
+
+## Updated modules
+- **[ADR Envelope](doc/Envelope.md#adr-envelope)**: Added inverted Envelope-output. Fixed issue where the high/low ouputs of BOA, EOA, BOR and EOR whould use gate-threshold values in stead of trigger-threshold values (only an issue if you changed output levels via the context menu). Also removed context menu-items to set gate-output levels, as the module don't have gate-outputs.
+- **[Bits-to-Value](doc/Bits.md#bits-to-value)**: Min/Center/Max mode-lights changed to tiny-lights.
+- **[Value-to-Bits](doc/Bits.md#value-to-bits)**: Min/Center/Max mode-lights changed to tiny-lights.
+
+## Fixed modules
+-
+
+
 # 2.0.3
 Ensured all module descriptions in manifest are identical to module descriptions in the manual. Also updated the manifest tags for several modules, e.g. LCMP2, VCMP1 and VCMP2 Mk I, are all able to work as 2-to-1 switches thanks to their "True"/"False" inputs in the top of the modules, but were not tagged as "Switch" in the maninfest. Various minor updates to the manual (e.g. added badges to highlight module features, and changed all module headers). Added support for multi-stage (color) push-buttons (needed one for the ADR Envelope).